requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==45.0.5
cycler==0.12.1
dlib==19.24.6
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
google-auth==2.40.3
google-genai==1.32.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
numpy==2.2.3
onnx==1.17.0
onnxruntime==1.21.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.1
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
